📝 Cargo metadata and 🎨 lib.rs
GH-16
